refactor: avoid one-word component names
Badge / Button / Card -> Base*
