+++ client/components/BaseCard.vue
