+++ client/pages/queue/[section].vue
@@ -52,7 +52,7 @@ import { DateTime } from 'luxon'
 import Fuse from 'fuse.js/basic'
 import { groupBy } from 'lodash-es'
 import { useSiteStore } from '@/stores/site'
-import Badge from '../../components/Badge.vue'
+import Badge from '../../components/BaseBadge.vue'
 import type { Column } from '~/components/DocumentTableTypes'
 import type { Assignment } from '~/rpctracker_client'
 import type { Tab } from '~/components/TabNavTypes'
